feat: improve note and directory ui, rename note-button component to index
