+++ .changeset/moody-books-fix.md
@@ -1,0 +1,5 @@
+---
+"ingenious": patch
+---
+
+feat: improve directory and note ui, rename note-button component
